[not verified] Moved the autoloader functions into a separate include.
