+++ .gitignore
@@ -57,3 +57,4 @@ _inc/jetpack-strings.php
 /vendor/composer/
 /vendor/autoload.php
 /vendor/autoload_packages.php
+/vendor/autoload_functions.php
