+++ packages/autoloader/src/AutoloadGenerator.php
@@ -38,6 +38,15 @@ use Composer\Util\Filesystem;
  */
 class AutoloadGenerator extends BaseGenerator {
 
+	const COMMENT = <<<AUTOLOADER_COMMENT
+/**
+ * This file was automatically generated by automattic/jetpack-autoloader.
+ *
+ * @package automattic/jetpack-autoloader
+ */
+
+AUTOLOADER_COMMENT;
+
 	/**
 	 * Instantiate an AutoloadGenerator object.
 	 *
@@ -89,8 +98,11 @@ class AutoloadGenerator extends BaseGenerator {
 		file_put_contents( $targetDir . '/jetpack_autoload_filemap.php', $this->getAutoloadFilesPackagesFile( $fileMap ) );
 		$this->io->writeError( '<info>Generated ' . $targetDir . '/jetpack_autoload_filemap.php</info>', true );
 
-		file_put_contents( $vendorPath . '/autoload_packages.php', $this->getAutoloadPackageFile( $suffix ) );
+		file_put_contents( $vendorPath . '/autoload_packages.php', $this->getAutoloadPackageFile( 'autoload.php', $suffix ) );
 		$this->io->writeError( '<info>Generated ' . $vendorPath . '/autoload_packages.php</info>', true );
+
+		file_put_contents( $vendorPath . '/autoload_functions.php', $this->getAutoloadPackageFile( 'functions.php', $suffix ) );
+		$this->io->writeError( '<info>Generated ' . $vendorPath . '/autoload_functions.php</info>', true );
 	}
 
 	/**
@@ -311,25 +323,23 @@ INCLUDE_FILEMAP;
 	/**
 	 * Generate the PHP that will be used in the autoload_packages.php files.
 	 *
-	 * @param String $suffix  Unique suffix added to the jetpack_enqueue_packages function.
+	 * @param String $filename a file to prepare.
+	 * @param String $suffix   Unique suffix used in the namespace.
 	 *
 	 * @return string
 	 */
-	private function getAutoloadPackageFile( $suffix ) {
-		$sourceLoader  = fopen( __DIR__ . '/autoload.php', 'r' );
+	private function getAutoloadPackageFile( $filename, $suffix ) {
+		$header  = self::COMMENT;
+		$header .= PHP_EOL . PHP_EOL;
+		$header .= 'namespace Automattic\Jetpack\Autoloader\jp' . $suffix . ';';
+		$header .= PHP_EOL . PHP_EOL;
+
+		$sourceLoader  = fopen( __DIR__ . '/' . $filename, 'r' );
 		$file_contents = stream_get_contents( $sourceLoader );
-		$file_contents = str_replace(
-			'Automattic\\Jetpack\\Autoloader;',
-			'Automattic\\Jetpack\\Autoloader\\jp' . $suffix . ';',
+		return str_replace(
+			'/* HEADER */',
+			$header,
 			$file_contents
 		);
-
-		$file_contents .= <<<INCLUDE_AUTOLOAD
-
-set_up_autoloader();
-
-INCLUDE_AUTOLOAD;
-
-		return $file_contents;
 	}
 }
